MAINT: move normalized_index_expression into normalize.py
